+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - fix method signature requirements(pr [#170](https://github.com/jerus-org/pcu/pull/170))
+- pr title elided (pr [#173])
 
 ## [0.1.8] - 2024-06-19
 
@@ -158,6 +159,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#167]: https://github.com/jerus-org/pcu/pull/167
 [#168]: https://github.com/jerus-org/pcu/pull/168
 [#171]: https://github.com/jerus-org/pcu/pull/171
+[#173]: https://github.com/jerus-org/pcu/pull/173
 [Unreleased]: https://github.com/jerus-org/pcu/compare/0.1.8...HEAD
 [0.1.8]: https://github.com/jerus-org/pcu/compare/0.1.7...0.1.8
 [0.1.7]: https://github.com/jerus-org/pcu/compare/0.1.6...0.1.7